requirements.txt
fastapi
uvicorn
pyorbital
python-dotenv